first file
I am a 20-year-old Korean male who just graduated from high school last year and is currently attending a large corporation.
My current job is very difficult and stressful for me. I made a little money designing and editing videos from an early age. Based on these experiences, I wanted to become a front-end developer, so I tried vs code for the first time today. It's still so hard I don't know what it is
Oh, and the picture that appears when you enter the link is my dog. If you'd like to use that photo or want to see more of my pup, send me a chat on Instagram.